# spec/openapi.yaml
openapi: 3.0.3
info:
  title: ZZAPI MES
  description: SAP ICF REST handlers for MES integration on SAP_BASIS 700
  version: 0.1.0
  license:
    name: UNLICENSED
servers:
  - url: http://sapdev.fastcell.hk:8000
    description: sapdev (Phase 1, direct SAP access)
  - url: http://localhost:8080
    description: zzapi-mes hub (Phase 3)
security: []
paths:
...
  /auth/token:
    post:
      operationId: createToken
      summary: Obtain a JWT bearer token
      description: Exchange an API key (format `<key_id>.<secret>`) for a short-lived JWT. The JWT carries `key_id`, `scopes`, `iat`, `exp`, and `rate_limit_per_min` claims.
      security: []  # No auth required for token issuance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties:
                api_key:
                  type: string
                  description: "API key in format: <key_id>.<secret>"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid, unknown, or revoked API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body too large (max 1 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Auth rate limit exceeded
          headers:
            retry-after:
              schema:
                type: integer
              description: Seconds until next request is allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /confirmation:
    post:
      operationId: createConfirmation
      summary: Production order confirmation
      description: Records a production order time ticket confirmation. Requires `conf` scope and an `Idempotency-Key` header for deduplication.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
          description: Client-generated idempotency key for deduplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmationRequest"
      responses:
        "201":
          description: Confirmation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmationResponse"
        "400":
          description: Invalid request body or missing Idempotency-Key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient scope (requires 'conf')
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body too large (max 1 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate Idempotency-Key, or SAP conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Idempotency-Key reused with different body, or SAP business rule rejection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            retry-after:
              schema:
                type: integer
              description: Seconds until next request is allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: SAP upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: SAP request timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goods-receipt:
    post:
      operationId: createGoodsReceipt
      summary: Goods receipt for purchase order
      description: Records a goods receipt (GR) against a purchase order. Requires `gr` scope and an `Idempotency-Key` header.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
          description: Client-generated idempotency key for deduplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoodsReceiptRequest"
      responses:
        "201":
          description: Goods receipt recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoodsReceiptResponse"
        "400":
          description: Invalid request body or missing Idempotency-Key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient scope (requires 'gr')
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body too large (max 1 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate Idempotency-Key, or SAP conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Idempotency-Key reused with different body, or SAP business rule rejection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            retry-after:
              schema:
                type: integer
              description: Seconds until next request is allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: SAP upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: SAP request timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goods-issue:
    post:
      operationId: createGoodsIssue
      summary: Goods issue for production order
      description: Records a goods issue (GI) for a production order component. Requires `gi` scope and an `Idempotency-Key` header.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
          description: Client-generated idempotency key for deduplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoodsIssueRequest"
      responses:
        "201":
          description: Goods issue recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoodsIssueResponse"
        "400":
          description: Invalid request body or missing Idempotency-Key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient scope (requires 'gi')
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body too large (max 1 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate Idempotency-Key, or SAP conflict, or SAP backflush conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Idempotency-Key reused with different body, or SAP business rule rejection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            retry-after:
              schema:
                type: integer
              description: Seconds until next request is allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: SAP upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: SAP request timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "PO not found"
    TokenResponse:
      type: object
      required: [token, expires_in]
      properties:
        token:
          type: string
          description: JWT bearer token with claims {key_id, scopes, iat, exp, rate_limit_per_min}
        expires_in:
          type: integer
          description: Token lifetime in seconds (default 900, configurable via HUB_JWT_TTL_SECONDS)
          example: 900
...
    ConfirmationRequest:
      type: object
      required: [orderid, operation, yield]
      properties:
        orderid:
          type: string
          minLength: 1
          maxLength: 12
          description: Production order number
          example: "1000000"
        operation:
          type: string
          minLength: 1
          maxLength: 4
          description: Operation number
          example: "0010"
        yield:
          type: number
          minimum: 1
          description: Yield quantity
          example: 50
        scrap:
          type: number
          minimum: 0
          description: Scrap quantity
          example: 0
        work_actual:
          type: number
          minimum: 0
          description: Actual work (hours)
          example: 1.5
        postg_date:
          type: string
          pattern: "^[0-9]{8}$"
          description: Posting date (YYYYMMDD)
          example: "20260422"
    ConfirmationResponse:
      type: object
      required: [orderid, operation, yield, scrap, status]
      properties:
        orderid:
          type: string
          description: Production order number
          example: "1000000"
        operation:
          type: string
          description: Operation number
          example: "0010"
        yield:
          type: number
          description: Yield quantity confirmed
          example: 50
        scrap:
          type: number
          description: Scrap quantity confirmed
          example: 0
        confNo:
          type: string
          description: Confirmation number from SAP
          example: "00000100"
        confCnt:
          type: string
          description: Confirmation counter from SAP
          example: "0001"
        status:
          type: string
          description: Confirmation status
          example: "confirmed"
        message:
          type: string
          description: Additional info
          example: "Production confirmation recorded"
    GoodsReceiptRequest:
      type: object
      required: [ebeln, ebelp, menge, werks, lgort]
      properties:
        ebeln:
          type: string
          minLength: 1
          maxLength: 10
          description: Purchasing document number
          example: "4500000001"
        ebelp:
          type: string
          minLength: 1
          maxLength: 5
          description: Item number
          example: "00010"
        menge:
          type: number
          minimum: 1
          description: Quantity to receive
          example: 100
        werks:
          type: string
          minLength: 1
          maxLength: 4
          description: Plant
          example: "1000"
        lgort:
          type: string
          minLength: 1
          maxLength: 4
          description: Storage location
          example: "0001"
        budat:
          type: string
          pattern: "^[0-9]{8}$"
          description: Posting date (YYYYMMDD)
          example: "20260422"
        charg:
          type: string
          maxLength: 10
          description: Batch number
          example: "BATCH001"
    GoodsReceiptResponse:
      type: object
      required: [ebeln, ebelp, menge, status]
      properties:
        ebeln:
          type: string
          description: Purchasing document number
          example: "4500000001"
        ebelp:
          type: string
          description: Item number
          example: "00010"
        menge:
          type: number
          description: Quantity received
          example: 100
        materialDocument:
          type: string
          description: Material document number
          example: "5000000001"
        documentYear:
          type: string
          description: Material document year
          example: "2026"
        status:
          type: string
          description: Receipt status
          example: "posted"
        message:
          type: string
          description: Additional info
          example: "Goods receipt posted"
    GoodsIssueRequest:
      type: object
      required: [orderid, matnr, menge, werks, lgort]
      properties:
        orderid:
          type: string
          minLength: 1
          maxLength: 12
          description: Production order number
          example: "1000000"
        matnr:
          type: string
          minLength: 1
          maxLength: 18
          description: Component material number
          example: "20000001"
        menge:
          type: number
          minimum: 1
          description: Quantity to issue
          example: 50
        werks:
          type: string
          minLength: 1
          maxLength: 4
          description: Plant
          example: "1000"
        lgort:
          type: string
          minLength: 1
          maxLength: 4
          description: Storage location
          example: "0001"
        budat:
          type: string
          pattern: "^[0-9]{8}$"
          description: Posting date (YYYYMMDD)
          example: "20260422"
        charg:
          type: string
          maxLength: 10
          description: Batch number
          example: "BATCH001"
    GoodsIssueResponse:
      type: object
      required: [orderid, matnr, menge, status]
      properties:
        orderid:
          type: string
          description: Production order number
          example: "1000000"
        matnr:
          type: string
          description: Component material number
          example: "20000001"
        menge:
          type: number
          description: Quantity issued
          example: 50
        materialDocument:
          type: string
          description: Material document number
          example: "5000000002"
        documentYear:
          type: string
          description: Material document year
          example: "2026"
        status:
          type: string
          description: Issue status
          example: "posted"
        message:
          type: string
          description: Additional info
          example: "Goods issue posted"